splited email domain into models and interfaces
