+++ src/email/application/email_service.py
@@ -1,5 +1,5 @@
-from ..domain.message import TemplateMessage
-from ..domain.message_sender_interface import MessageSenderInterface
+from ..domain.models.message import TemplateMessage
+from ..domain.interfaces.message_sender_interface import MessageSenderInterface
 
 
 class EmailService:
