+++ src/email/domain/interfaces/message_sender_interface.py
@@ -1,5 +1,5 @@
 from abc import ABC, abstractmethod
-from ..domain.message import TemplateMessage
+from ..models.message import TemplateMessage
 
 class MessageSenderInterface(ABC):
     @abstractmethod
