+++ src/email/domain/interfaces/templater_interface.py
@@ -1,0 +1,7 @@
+from abc import ABC, abstractmethod
+
+
+class TemplaterInterface(ABC):
+    @abstractmethod
+    def render(template_name: str, context: dict) -> str:
+        ...
